+++ backend/models/message_event.py
@@ -5,6 +5,11 @@ from pydantic import BaseModel
 from models.conversation import Conversation, Message, ConversationPhoto
 
 
+# Freemium action constants
+FREEMIUM_ACTION_SETUP_ON_DEVICE_STT = "setup_on_device_stt"
+FREEMIUM_ACTION_NONE = "none"
+
+
 class MessageEvent(BaseModel):
     event_type: str
 
@@ -148,6 +153,18 @@ class SpeakerLabelSuggestionEvent(MessageEvent):
     person_id: str
     person_name: str
     segment_id: str
+
+    def to_json(self):
+        j = self.model_dump(mode="json")
+        j["type"] = self.event_type
+        del j["event_type"]
+        return j
+
+
+class FreemiumThresholdReachedEvent(MessageEvent):
+    event_type: str = "freemium_threshold_reached"
+    remaining_seconds: int
+    action: str
 
     def to_json(self):
         j = self.model_dump(mode="json")
